Implemented Cache

+++ team9_project3.go
@@ -31,7 +31,7 @@ var InstructionList []Instruction
 // holds registers R0 - R31 (default 0)
 var Registers [32]int
 
-var Mem = make(map[int]int64)
+var Mem = make(map[int]int)
 
 func main() {
 	inputFilePathPtr := flag.String("i", "executionTest.txt", "input file path")
